settings: Look for gadget templates in default-template-path
Gadget templates are searched for in all paths of lookup-path. However,
gt needs to know which of them to use to save templates. That is why
there is a separate entry in this file to specify that, and it is called
default-template-path.

With current configuration it is impossible to save and then load the saved
templates, because they will be searched for in a different directory.

This patch sets the first path of lookup-path to point to the same place
as default-template-path.

Signed-off-by: Andrzej Pietrasiewicz <andrzej.p@collabora.com>