settings: make hardcoded defaults reflect suggested defaults
gt.conf installed with gt contains suggested default values, which are in
fact commented out.

This patch makes gt follow this convention: if after a fresh install an
entry in config file is commented out, the user can assume that the entry
is actually in effect. In other words: if user does not modify the config
file, then the install-time comments in it suggest what values are assumed.

Signed-off-by: Andrzej Pietrasiewicz <andrzej.p@collabora.com>