cmake: detect asciidoc a2x and disable the manual page when it is missing
Support forcing manual page building on or off so that builders can
fail builds without manual pages or silence the warning if needed.

This is useful in situations where gt is being installed on devices that
don't have manual pages or don't have a way to read them and so that the
person building gt doesn't have to manually disable the manual page.

Co-authored-by: Sergei Nikulov <sergey.nikulov@gmail.com>
Closes: https://github.com/linux-usb-gadgets/gt/pull/25